web_gateway/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
grpcio==1.67.1
grpcio-tools==1.67.1
